editor/composer/test/test_bug1200533.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1200533
-->
<head>
  <title>Test for Bug 1200533</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1200533">Mozilla Bug 1200533</a>
<p id="display"></p>
<iframe id="content"></iframe>

</div>
<pre id="test">
<script class="testbody" ttype="application/javascript">

/** Test for Bug 1200533 **/
/** Visit the elements defined above and check the dictionary we got **/
SimpleTest.waitForExplicitFinish();
var content = document.getElementById('content');

var tests = [
    // text area, value of spellchecker.dictionary, result.
    // Result: Document language.
    [ "none",  "", "en-US" ],
    // Result: Element language.
    [ "en-GB", "", "en-GB" ],
    [ "en-gb", "", "en-GB" ],
    // Result: Random en-*.
    [ "en-ZA-not-avail", "", "*" ],
    [ "en-generic",      "", "*" ],
    // Result: Locale.
    [ "ko-not-avail",    "", "en-US" ],

    // Result: Preference value in all cases.
    [ "en-ZA-not-avail", "en-AU", "en-AU" ],
    [ "en-generic",      "en-AU", "en-AU" ],
    [ "ko-not-avail",    "en-AU", "en-AU" ],

    // Result: Random en-*.
    [ "en-ZA-not-avail", "de-DE", "*" ],
    [ "en-generic",      "de-DE", "*" ],
    // Result: Preference value.
    [ "ko-not-avail",    "de-DE", "de-DE" ],
  ];

var loadCount = 0;
var en_GB;
var en_AU;
var en_DE;
var hunspell;

var loadListener = function(evt) {
  Components.utils.import("resource://gre/modules/AsyncSpellCheckTestHelper.jsm");
  Components.utils.import("resource://gre/modules/Services.jsm");

  if (loadCount == 0) {
    var dir = Components.classes["@mozilla.org/file/directory_service;1"]
                        .getService(Components.interfaces.nsIProperties)
                        .get("CurWorkD", Components.interfaces.nsIFile);
    dir.append("tests");
    dir.append("editor");
    dir.append("composer");
    dir.append("test");

    hunspell = Components.classes["@mozilla.org/spellchecker/engine;1"]
                         .getService(Components.interfaces.mozISpellCheckingEngine);

    // Install en-GB, en-AU and de-DE dictionaries.
    en_GB = dir.clone();
    en_AU = dir.clone();
    de_DE = dir.clone();
    en_GB.append("en-GB");
    en_AU.append("en-AU");
    de_DE.append("de-DE");
    is(en_GB.exists(), true, "true expected (en-GB directory should exist)");
    is(en_AU.exists(), true, "true expected (en-AU directory should exist)");
    is(de_DE.exists(), true, "true expected (de-DE directory should exist)");
    hunspell.addDirectory(en_GB);
    hunspell.addDirectory(en_AU);
    hunspell.addDirectory(de_DE);
  }

  Services.prefs.setCharPref("spellchecker.dictionary", tests[loadCount][1]);

  var doc = evt.target.contentDocument;
  var elem = doc.getElementById(tests[loadCount][0]);
  var editor = elem.QueryInterface(Components.interfaces.nsIDOMNSEditableElement).editor;
  editor.setSpellcheckUserOverride(true);
  var inlineSpellChecker = editor.getInlineSpellChecker(true);

  onSpellCheck(elem, function () {
    var spellchecker = inlineSpellChecker.spellChecker;
    try {
      var dict = spellchecker.GetCurrentDictionary();
    } catch(e) {}

    if (tests[loadCount][2] != "*") {
      is (dict, tests[loadCount][2], "expected " + tests[loadCount][2]);
    } else {
      var gotEn = (dict == "en-GB" || dict == "en-AU" || dict == "en-US");
      is (gotEn, true, "expected en-AU or en-GB or en-US");
    }

    loadCount++;
    if (loadCount < tests.length) {
      // Load the iframe again.
      content.src = 'http://mochi.test:8888/tests/editor/composer/test/bug1200533_subframe.html?firstload=false';
    } else {
      // Remove the fake  dictionaries again, since it's otherwise picked up by later tests.
      hunspell.removeDirectory(en_GB);
      hunspell.removeDirectory(en_AU);
      hunspell.removeDirectory(de_DE);

      // Reset the preference, so the last value we set doesn't collide with the next test.
      Services.prefs.setCharPref("spellchecker.dictionary", "");

      SimpleTest.finish();
    }
  });

}

content.addEventListener('load', loadListener, false);

content.src = 'http://mochi.test:8888/tests/editor/composer/test/bug1200533_subframe.html?firstload=true';

</script>
</pre>
</body>
</html>
